requirements.txt
alembic==1.12.1
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.14
decorator==5.2.1
docker==6.1.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.12.4
Flask==3.0.0
fnmatch2==0.0.8
fonttools==4.43.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.2.1
gunicorn==21.2.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
Mako==1.2.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.7
mlflow==2.8.0
monai==1.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.4
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==13.0.0
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.4.0
querystring-parser==1.2.4
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.16.0
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.13
websocket-client==1.6.4
Werkzeug==3.0.1
zipp==3.17.0
